Un-publish 2016 new years resolution
This blog post is pretty embarrassing because of how quickly and totally
I failed at it. In the interest of shamefully hiding my mistakes, I am
going to retroactively remove this blog post. Surely it's been caught by
the Internet Archive by now so I won't be able to get away with it
forever.